pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "torch>=2.0.1",
    "sentence-transformers>=2.2.2",
    "ipython>=8.13.2",
    "sourcery-cli>=1.2.0",
    "numpy>=1.24.3",
    "polars>=0.17.13",
    "seaborn>=0.12.2",
    "wordcloud>=1.9.1.1",
    "bertopic>=0.14.1",
    "pyarrow>=12.0.0",
    "pyqt5>=5.15.9",
    "jupyter>=1.0.0",
    "geopandas>=0.13.0",
    "tabulate>=0.9.0",
    "ipdb>=0.13.13",
    "datasets>=2.12.0",
    "scienceplots>=2.0.1",
    "dvc>=2.58.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
